Move utils/test to testing/helpers
